Лайки и дизлйки для Note + правки

+++ src/main/java/com/dev/GymForDevelopers/models/DTO/GdCommentDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Data
 @Builder(toBuilder = true)
 public class GdCommentDTO {
-    private Integer noteId;
+    private Long noteId;
 
     private String author;
