+++ src/main/java/com/dev/GymForDevelopers/services/GdCommentService.java
@@ -8,7 +8,6 @@ import com.dev.GymForDevelopers.repositories.GdCommentRepository;
 import com.dev.GymForDevelopers.repositories.GdNoteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 
@@ -35,7 +34,7 @@ public class GdCommentService {
      * @param gdComment комментарий
      */
     //todo поле author будет браться из токена
-    public GdComment create(Integer noteId, GdComment gdComment) {
+    public GdComment create(Long noteId, GdComment gdComment) {
         if (gdComment == null) {
             throw new GdRuntimeException("В качестве комментария был передан null", ExceptionConst.ERRORS_CODE_RT);
         }
